refactor filters subpackage into a single module to be consistent with DRF base
