--- rest_framework_json_api/filters/__init__.py
@@ -1,2 +1,0 @@
-from .sort import OrderingFilter  # noqa: F401
-from .queryvalidation import QueryParameterValidationFilter  # noqa: F401
